Add public_controller config for public routes
Introduces a separate base controller for public routes (like invitation
acceptance) that don't require authentication.

The problem: Host apps often add authentication filters to their
ApplicationController. Since engine controllers inherit from the host's
ApplicationController, these filters interfere with public routes like
viewing/accepting invitations.

The solution: Add a `public_controller` config option (defaults to
ActionController::Base) and a new PublicController base class. Public
routes now use PublicInvitationsController which inherits from this
minimal base, avoiding host app filter pollution.

Changes:
- Add config.public_controller option (default: ActionController::Base)
- Add Organizations::PublicController base class
- Add Organizations::PublicInvitationsController for show/accept actions
- Update routes to use new controller for invitation acceptance
- Clean up InvitationsController (remove public actions)
- Update README with new config option
